CTD-619 problem with auto- install.sh (#144)
* Update auto-install.sh

Added CONNECTDVERSION to reference the specific release of connectd daemon used by this package.  Prevents issue when "latest" release changes the name of one of the daemons.

* Update auto-install.sh

Removed "downloadSchannel" and "downMultiport" functions and references as these were never used.  Only the connectd daemon download is used for explicit testing.

* Update auto-install.sh

Updated comment to refer to connectd only as we no longer download the schannel or mux/demux binaries (they are not being used).